+++ packages/web/src/lib/view/parts-list/shared.ts
